rabbitmq-exporter/requirements.txt
prometheus_client==0.17.1
requests==2.31.0
python-dotenv==1.0.0 